Extract `edit_market` tests
